+++ site/dat/docs/Changelog.md
@@ -11,6 +11,7 @@
  - add `public-report` option to BlazeMeter reporter and cloud provisioning
  - add `retrieve-resources-regex` scenario-level option to JMeter
  - fix latency growing problem
+ - allow controling JMeter HTTP redirects with `follow-redirects` option
 
 ## 1.7.4 <sup>11 nov 2016</sup>
  - fix Locust crash when used with 'requests'-style scenario and cloud provisioning